requirements.txt
matplotlib==3.10.8
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
scikit-learn==1.8.0
seaborn==0.13.2
torch>=2.0.0
torchvision>=0.15.0
tqdm==4.67.1
transformers==4.57.3
